engine: complete the layer manifest from the checkpoint
My earlier tensor list to this lane was incomplete: KDA has separate q/k/v
projections and q/k/v conv1d, and the norms, convs and the two KDA scalars are
F32 while every projection matrix is BF16. The manifest now carries the full
layer-12 table read from the safetensors header, dtypes included, and the shapes
test fails in both directions - a parameter the manifest lacks, or a manifest
entry the implementation never loads.

Gated: 474 passed.

+++ engine/k3ref/router.py
@@ -37,7 +37,7 @@ class K3Router(nn.Module):
             torch.empty(num_experts, hidden_size, device=device, dtype=dtype)
         )
         self.e_score_correction_bias = nn.Parameter(
-            torch.empty(num_experts, device=device, dtype=dtype)
+            torch.empty(num_experts, device=device, dtype=torch.float32)
         )
         nn.init.kaiming_uniform_(self.weight, a=5**0.5)
         nn.init.zeros_(self.e_score_correction_bias)
